+++ apps/web/src/pages/Auth/RegistrationPages/utils/messages.ts
@@ -1,0 +1,21 @@
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  breadcrumb: {
+    defaultMessage: "Registration",
+    id: "H6FgWU",
+    description: "Breadcrumb for the registration pages",
+  },
+  title: {
+    defaultMessage: "Welcome to GC Digital Talent",
+    id: "D0cHjz",
+    description: "Page title for the registration pages",
+  },
+  subtitle: {
+    defaultMessage: "Get started by completing your basic account information.",
+    id: "g+tCWe",
+    description: "Subtitle for the registration pages",
+  },
+});
+
+export default messages;
